refactor: move render to search.ts
